🔧 Project structure refactor: organized frontend/backend/contracts folders and fixed imports
