+++ astro.config.mjs
@@ -1,15 +1,18 @@
 // @ts-check
 import { defineConfig } from 'astro/config';
 import tailwindcss from '@tailwindcss/vite';
-
 import vercel from '@astrojs/vercel';
+import dotenv from 'dotenv'; // ✅ تحميل .env
+
+dotenv.config(); // ✅ تحميل متغيرات البيئة
 
 // https://astro.build/config
 export default defineConfig({
-  output: "server",
+  // ✅ المهم: نقول ل Astro إن ملفاتنا في frontend/src بدل src
+  srcDir: 'frontend/src',
+  output: 'server',
   vite: {
-      plugins: [tailwindcss()]
-	},
-
+    plugins: [tailwindcss()]
+  },
   adapter: vercel()
 });
