+++ frontend/src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-// استدعاء ملف الستايلات
+
 import "../styles/global.css";
 ---
