+++ package.json
@@ -21,6 +21,7 @@
     "astro": "^5.8.0",
     "base64url": "^3.0.1",
     "canvas-confetti": "^1.9.3",
+    "dotenv": "^16.5.0",
     "elysia": "^1.3.1",
     "tailwindcss": "^4.1.3"
   }
